More classes and file management
Added the class Documents, which stores the words of the documents in the desired format, and the class Map, that doesnt do anything yet but will combine the classes Vector and Document in the future. Also made some changes to the main(), that now is able to read its first file and convert it into a Document.